+++ handler/api/repos/enable.go
@@ -16,6 +16,7 @@ package repos
 
 import (
 	"net/http"
+	"os"
 
 	"github.com/drone/drone/core"
 	"github.com/drone/drone/handler/api/render"
@@ -25,6 +26,12 @@ import (
 	"github.com/dchest/uniuri"
 	"github.com/go-chi/chi"
 )
+
+// FEATURE FLAG enables a static secret value used to sign
+// incoming requests routed through a proxy. This was implemented
+// based on feedback from @chiraggadasc and and should not be
+// removed until we have a permanent solution in place.
+var staticSigner = os.Getenv("DRONE_FEATURE_SERVER_PROXY_SECRET")
 
 // HandleEnable returns an http.HandlerFunc that processes http
 // requests to enable a repository in the system.
@@ -63,6 +70,10 @@ func HandleEnable(
 		}
 		if repo.Timeout == 0 {
 			repo.Timeout = 60
+		}
+
+		if staticSigner != "" {
+			repo.Signer = staticSigner
 		}
 
 		err = hooks.Create(r.Context(), user, repo)